backend/requirements.txt
# LiveKit Agents Framework
livekit-agents[openai]
livekit-plugins-openai
# livekit-plugins-silero  # Commented out due to Windows onnxruntime DLL issues
livekit-api
openai[realtime]
azure-identity

# Web Framework for Token Server
flask
flask-cors

# Environment Management
python-dotenv

# Optional: For enhanced audio processing (if needed)
# numpy
# sounddevice
